Move subslice pattern tests to alloc/tests/slice.rs
